Alter command's argument processing to match Unix norms.
Use stderr for all diagnostics rather than mixing with stdout.
Exit with a non-zero status on error.
Detail the nature of the error, e.g. user may not realise they're given
two arguments due to a space in the path.
Use the OS-independent basename of the executable in the error message.
Turn off log's default prefix of date and time for all log messages.