Make WorkingTree into an interface
New type anyWorkingTree combines a GoSource with a vcs.Cmd, and provides
a pass-through methods for TagSync.

WorkingTree is an interface implemented for git by gitWorkingTree.

NewWorkingTree now returns an interface, the underlying value of which
is determined by the vcs in use.

NewHasher returns an interface, the underlying value of which is
determined by the vcs in use.

PseudoVersion is now a function which acts on a Describable, that being
the interface which is the subset of methods needed to create a
pseudo-version.

Signed-off-by: Tim Waugh <twaugh@redhat.com>